requirements.txt
python-telegram-bot[job-queue]==21.10
python-dotenv==1.0.1
